requirements.txt
asgiref==3.7.2
Django==4.2
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
